requirements.txt
pytest==8.3.5
allure-python-commons==2.13.5
requests==2.32.3
PyYAML==6.0.2